requirements.txt
python
sklearn
xgboost
catboost
pandas
numpy
seaborn
matplotlib
pymongo
python-dotenv
cassandra-driver

